pyproject.toml
[tool.poetry]
name = "chat-gpt"
version = "0.1.0"
description = ""
authors = ["fastcode <fast.code.auth@gmail.com>"]
readme = "README.md"
packages = [{include = "chat_gpt"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
aiogram = "3.0.0b6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
